Dissolved grakn.core.graql.util
- Moved StringUtils method into StringPrinter
- Moved PrimitiveNumberComparator as a nested class in AggregateExecutor
