--- server/src/graql/util/StringUtil.java
@@ -1,51 +1,0 @@
-/*
- * GRAKN.AI - THE KNOWLEDGE GRAPH
- * Copyright (C) 2018 Grakn Labs Ltd
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package grakn.core.graql.util;
-
-import grakn.core.graql.concept.ConceptId;
-import grakn.core.graql.concept.Label;
-
-import static graql.lang.util.StringUtil.escapeLabelOrId;
-
-/**
- * Some helper methods in dealing with strings in the context of Graql.
- */
-public class StringUtil {
-
-    /**
-     * @param id an ID of a concept
-     * @return
-     * The id of the concept correctly escaped in graql.
-     * If the ID doesn't begin with a number and is only comprised of alphanumeric characters, underscores and dashes,
-     * then it will be returned as-is, otherwise it will be quoted and escaped.
-     */
-    public static String idToString(ConceptId id) {
-        return escapeLabelOrId(id.getValue());
-    }
-
-    /**
-     * @param label a label of a type
-     * @return
-     * The label of the type correctly escaped in graql.
-     * If the label doesn't begin with a number and is only comprised of alphanumeric characters, underscores and dashes,
-     * then it will be returned as-is, otherwise it will be quoted and escaped.
-     */
-    public static String typeLabelToString(Label label) {
-        return escapeLabelOrId(label.getValue());
-    }
-}
